+++ CHANGELOG.md
@@ -180,6 +180,8 @@ Baseline (changes not logged before this point in time).
 
 ### V. 0.4.2 (next version number)
 
+* Limited the number of threads (1023) and processes (127) that can be invoked without changing the library's code.
+
 * Minor performance oriented changes.
 
 * Fixed an issue where Websocket upgrade would allow code execution in parallel with `on_open` (protocol locking was fixed while switching the protocol).